v4/Audio-Transcription/requirements.txt
whisperx @ git+https://github.com/m-bain/whisperX.git@6ec4a020489d904c4f2cd1ed097674232d2692d4
--extra-index-url https://download.pytorch.org/whl/cpu
transformers==4.57.6
faster-whisper==1.2.1
ctranslate2==4.7.1
numpy==2.2.6
pyannote.audio==3.4.0

fastapi==0.128.1
uvicorn==0.40.0
python-dotenv==1.2.1
psutil==7.2.2
boto3==1.42.42
requests==2.32.5
python-multipart==0.0.22
Jinja2==3.1.6
pydub==0.25.1
openai==2.16.0


pytorch-lightning==2.4.0

#Install these seprately 
#pip install torch==2.5.1+cpu torchaudio==2.5.1+cpu torchvision==0.20.1+cpu   --index-url https://download.pytorch.org/whl/cpu
